Se modifico las consultas, migraciones, seeders, email de creacion y email de password, se crea rule para el case-sensitibity del unique de posgres y se modifico los request con esta regla

+++ database/migrations/2019_08_29_145814_create_trm_table.php
@@ -15,7 +15,7 @@ class CreateTrmTable extends Migration
     {
         Schema::create('trm', function (Blueprint $table) {
             $table->bigIncrements('id');
-						$table->double('value');
+						$table->float('value');
 						$table->boolean('status')->default(1);
             $table->timestamps();
         });
